# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Lightweight proxy core management panel for Xray, Sing-box, and Mihomo.",
        "title": "Isolate Panel API",
        "contact": {
            "name": "Isolate Panel",
            "url": "https://github.com/isolate-project/isolate-panel/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/inbounds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all inbounds, optionally filtered by core_id or is_enabled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "List inbounds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by core ID",
                        "name": "core_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by enabled state",
                        "name": "is_enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new proxy inbound (listener) on a core",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Create inbound",
                "parameters": [
                    {
                        "description": "Inbound configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_isolate-project_isolate-panel_internal_models.Inbound"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grant a user access to an inbound",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Assign inbound to user",
                "parameters": [
                    {
                        "description": "{user_id, inbound_id}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/check-port": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns whether the given port is available or already in use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Check port availability",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Port number to check",
                        "name": "port",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Inbound ID to exclude from conflict check",
                        "name": "exclude_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/core/{core_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all inbounds belonging to a specific core",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Get inbounds by core",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Core ID",
                        "name": "core_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/unassign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke a user's access to an inbound",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Unassign inbound from user",
                "parameters": [
                    {
                        "description": "{user_id, inbound_id}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single inbound by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Get inbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update inbound fields (partial update via map)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Update inbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an inbound and remove all user assignments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Delete inbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/{id}/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all users that have access to a specific inbound",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Get inbound users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/inbounds/{id}/users/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add or remove multiple users from an inbound in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inbounds"
                ],
                "summary": "Bulk assign users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "{add_user_ids: [], remove_user_ids: []}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_isolate-project_isolate-panel_internal_models.Core": {
            "type": "object",
            "properties": {
                "config_path": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "is_running": {
                    "type": "boolean"
                },
                "last_error": {
                    "type": "string"
                },
                "log_path": {
                    "type": "string"
                },
                "name": {
                    "description": "singbox, xray, mihomo",
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "restart_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "uptime_seconds": {
                    "description": "Statistics",
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "github_com_isolate-project_isolate-panel_internal_models.Inbound": {
            "type": "object",
            "properties": {
                "config_json": {
                    "type": "string"
                },
                "core": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_isolate-project_isolate-panel_internal_models.Core"
                        }
                    ]
                },
                "core_id": {
                    "type": "integer"
                },
                "created_at": {
                    "description": "Metadata",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "description": "Status",
                    "type": "boolean"
                },
                "listen_address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "reality_config_json": {
                    "type": "string"
                },
                "reality_enabled": {
                    "type": "boolean"
                },
                "tls_cert_id": {
                    "type": "integer"
                },
                "tls_enabled": {
                    "description": "TLS/REALITY",
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}